[ai-bot refactor] Move handleDebug command out + misc (#1254)
* move debug into a module

* different handler for debug commands

* introduce assistant class

* move event sending out of main

* remove circular import for matrix

* Do not need history if you are setting title via debug

* add order and room message commands

* optionally disable matrix js logging

* cleanup logs

* cleanup

* cleanup
