+++ packages/ai-bot/main.ts
@@ -1,11 +1,11 @@
 import './setup-logger'; // This should be first
 import {
-  IContent,
   RoomMemberEvent,
   RoomEvent,
   createClient,
   Room,
   MatrixClient,
+  type MatrixEvent,
 } from 'matrix-js-sdk';
 import OpenAI from 'openai';
 import { logger, aiBotUsername } from '@cardstack/runtime-common';
@@ -17,6 +17,8 @@ import {
   isPatchReactionEvent,
 } from './helpers';
 import { shouldSetRoomTitle, setTitle } from './lib/set-title';
+import { handleDebugCommands } from './lib/debug';
+import { sendError, sendOption, sendMessage } from './lib/matrix';
 import { OpenAIError } from 'openai/error';
 import type { MatrixEvent as DiscreteMatrixEvent } from 'https://cardstack.com/base/room';
 import * as Sentry from '@sentry/node';
@@ -30,210 +32,55 @@ if (process.env.SENTRY_DSN) {
 
 let log = logger('ai-bot');
 
-const openai = new OpenAI();
+class Assistant {
+  private openai: OpenAI;
+  private client: MatrixClient;
+  id: string;
+
+  constructor(client: MatrixClient, id: string) {
+    this.openai = new OpenAI();
+    this.id = id;
+    this.client = client;
+  }
+
+  getResponse(history: DiscreteMatrixEvent[]) {
+    let tools = getTools(history, this.id);
+    let messages = getModifyPrompt(history, this.id, tools);
+    if (tools.length === 0) {
+      return this.openai.beta.chat.completions.stream({
+        model: 'gpt-4o',
+        messages: messages,
+      });
+    } else {
+      return this.openai.beta.chat.completions.stream({
+        model: 'gpt-4o',
+        messages: messages,
+        tools: tools,
+        tool_choice: 'auto',
+      });
+    }
+  }
+
+  async handleDebugCommands(eventBody: string, room: Room) {
+    return handleDebugCommands(
+      this.openai,
+      eventBody,
+      this.client,
+      room,
+      this.id,
+    );
+  }
+
+  async setTitle(
+    room: Room,
+    history: DiscreteMatrixEvent[],
+    event?: MatrixEvent,
+  ) {
+    return setTitle(this.openai, this.client, room, history, this.id, event);
+  }
+}
 
 let startTime = Date.now();
-
-async function sendEvent(
-  client: MatrixClient,
-  room: Room,
-  eventType: string,
-  content: IContent,
-  eventToUpdate: string | undefined,
-) {
-  if (content.data) {
-    content.data = JSON.stringify(content.data);
-  }
-  if (eventToUpdate) {
-    content['m.relates_to'] = {
-      rel_type: 'm.replace',
-      event_id: eventToUpdate,
-    };
-  }
-  log.info('Sending', content);
-  return await client.sendEvent(room.roomId, eventType, content);
-}
-
-async function sendMessage(
-  client: MatrixClient,
-  room: Room,
-  content: string,
-  eventToUpdate: string | undefined,
-  data: any = {},
-) {
-  log.info('Sending', content);
-  let messageObject: IContent = {
-    ...{
-      body: content,
-      msgtype: 'm.text',
-      formatted_body: content,
-      format: 'org.matrix.custom.html',
-      'm.new_content': {
-        body: content,
-        msgtype: 'm.text',
-        formatted_body: content,
-        format: 'org.matrix.custom.html',
-      },
-    },
-    ...data,
-  };
-  return await sendEvent(
-    client,
-    room,
-    'm.room.message',
-    messageObject,
-    eventToUpdate,
-  );
-}
-
-// TODO we might want to think about how to handle patches that are larger than
-// 65KB (the maximum matrix event size), such that we split them into fragments
-// like we split cards into fragments
-async function sendOption(
-  client: MatrixClient,
-  room: Room,
-  patch: any,
-  eventToUpdate: string | undefined,
-) {
-  log.info('sending option', patch);
-  const id = patch['card_id'];
-  const body = patch['description'] || "Here's the change:";
-  let messageObject = {
-    body: body,
-    msgtype: 'org.boxel.command',
-    formatted_body: body,
-    format: 'org.matrix.custom.html',
-    data: {
-      command: {
-        type: 'patch',
-        id: id,
-        patch: {
-          attributes: patch['attributes'],
-          relationships: patch['relationships'],
-        },
-        eventId: eventToUpdate,
-      },
-    },
-  };
-  log.info(JSON.stringify(messageObject, null, 2));
-  return await sendEvent(
-    client,
-    room,
-    'm.room.message',
-    messageObject,
-    eventToUpdate,
-  );
-}
-
-function getResponse(history: DiscreteMatrixEvent[], aiBotUsername: string) {
-  let tools = getTools(history, aiBotUsername);
-  let messages = getModifyPrompt(history, aiBotUsername, tools);
-  if (tools.length === 0) {
-    return openai.beta.chat.completions.stream({
-      model: 'gpt-4o',
-      messages: messages,
-    });
-  } else {
-    return openai.beta.chat.completions.stream({
-      model: 'gpt-4o',
-      messages: messages,
-      tools: tools,
-      tool_choice: 'auto',
-    });
-  }
-}
-
-function getErrorMessage(error: any): string {
-  if (error instanceof OpenAIError) {
-    return `OpenAI error: ${error.name} - ${error.message}`;
-  }
-  if (typeof error === 'string') {
-    return `Unknown error: ${error}`;
-  }
-  return `Unknown error`;
-}
-
-async function sendError(
-  client: MatrixClient,
-  room: Room,
-  error: any,
-  eventToUpdate: string | undefined,
-) {
-  try {
-    let errorMessage = getErrorMessage(error);
-    log.error(errorMessage);
-    await sendMessage(
-      client,
-      room,
-      'There was an error processing your request, please try again later',
-      eventToUpdate,
-      {
-        isStreamingFinished: true,
-        errorMessage,
-      },
-    );
-  } catch (e) {
-    // We've had a problem sending the error message back to the user
-    // Log and continue
-    log.error(`Error sending error message back to user: ${e}`);
-    Sentry.captureException(e);
-  }
-}
-
-async function handleDebugCommands(
-  eventBody: string,
-  client: MatrixClient,
-  room: Room,
-  history: DiscreteMatrixEvent[],
-  userId: string,
-) {
-  // Explicitly set the room name
-  if (eventBody.startsWith('debug:title:set:')) {
-    return await client.setRoomName(
-      room.roomId,
-      eventBody.split('debug:title:set:')[1],
-    );
-  } else if (eventBody.startsWith('debug:boom')) {
-    await sendError(
-      client,
-      room,
-      `Boom! Throwing an unhandled error`,
-      undefined,
-    );
-    throw new Error('Boom!');
-  }
-  // Use GPT to set the room title
-  else if (eventBody.startsWith('debug:title:create')) {
-    return await setTitle(openai, client, room, history, userId);
-  } else if (eventBody.startsWith('debug:patch:')) {
-    let patchMessage = eventBody.split('debug:patch:')[1];
-    // If there's a card attached, we need to split it off to parse the json
-    patchMessage = patchMessage.split('(Card')[0];
-    let command: {
-      card_id?: string;
-      description?: string;
-      attributes?: any;
-    } = {};
-    try {
-      command = JSON.parse(patchMessage);
-      if (!command.card_id || !command.description || !command.attributes) {
-        throw new Error(
-          'Invalid debug patch: card_id, description, or attributes is missing.',
-        );
-      }
-    } catch (error) {
-      Sentry.captureException(error);
-      return await sendMessage(
-        client,
-        room,
-        `Error parsing your debug patch, ${error} ${patchMessage}`,
-        undefined,
-      );
-    }
-    return await sendOption(client, room, command, undefined);
-  }
-  return;
-}
 
 (async () => {
   const matrixUrl = process.env.MATRIX_URL || 'http://localhost:8008';
@@ -264,12 +111,14 @@ Common issues are:
       process.exit(1);
     });
   let { user_id: aiBotUserId } = auth;
+  let assistant = new Assistant(client, aiBotUserId);
+
   client.on(RoomMemberEvent.Membership, function (_event, member) {
     if (member.membership === 'invite' && member.userId === aiBotUserId) {
       client
         .joinRoom(member.roomId)
         .then(function () {
-          log.info('Auto-joined %s', member.roomId);
+          log.info('%s auto-joined %s', member.name, member.roomId);
         })
         .catch(function (err) {
           log.info(
@@ -289,7 +138,6 @@ Common issues are:
         if (!room) {
           return;
         }
-        log.info('(%s) %s :: %s', room?.name, event.getSender(), eventBody);
 
         if (event.event.origin_server_ts! < startTime) {
           return;
@@ -306,26 +154,21 @@ Common issues are:
         if (event.getSender() === aiBotUserId) {
           return;
         }
+        log.info(
+          '(%s) (Room: "%s" %s) (Message: %s %s)',
+          event.getType(),
+          room?.name,
+          room?.roomId,
+          event.getSender(),
+          eventBody,
+        );
 
         let initial = await client.roomInitialSync(room!.roomId, 1000);
         let eventList = (initial!.messages?.chunk ||
           []) as DiscreteMatrixEvent[];
-        log.info(eventList);
-
         log.info('Total event list', eventList.length);
         let history: DiscreteMatrixEvent[] = constructHistory(eventList);
         log.info("Compressed into just the history that's ", history.length);
-
-        // To assist debugging, handle explicit commands
-        if (eventBody.startsWith('debug:')) {
-          return await handleDebugCommands(
-            eventBody,
-            client,
-            room,
-            history,
-            aiBotUserId,
-          );
-        }
 
         let initialMessage = await sendMessage(
           client,
@@ -336,7 +179,8 @@ Common issues are:
 
         let unsent = 0;
         let thinkingMessageReplaced = false;
-        const runner = getResponse(history, aiBotUserId)
+        const runner = assistant
+          .getResponse(history)
           .on('content', async (_delta, snapshot) => {
             unsent += 1;
             if (unsent > 40) {
@@ -356,7 +200,7 @@ Common issues are:
             if (msg.role === 'assistant') {
               for (const toolCall of msg.tool_calls || []) {
                 const functionCall = toolCall.function;
-                console.log('Function call', toolCall);
+                log.debug('[Room Timeline] Function call', toolCall);
                 let args;
                 try {
                   args = JSON.parse(functionCall.arguments);
@@ -412,14 +256,7 @@ Common issues are:
         }
 
         if (shouldSetRoomTitle(eventList, aiBotUserId, event)) {
-          return await setTitle(
-            openai,
-            client,
-            room,
-            history,
-            aiBotUserId,
-            event,
-          );
+          return await assistant.setTitle(room, history, event);
         }
         return;
       } catch (e) {
@@ -430,33 +267,59 @@ Common issues are:
     },
   );
 
+  //handle reaction events
   client.on(RoomEvent.Timeline, async function (event, room) {
-    if (event.getType() !== 'm.reaction') {
+    if (!room) {
       return;
     }
+    if (!isPatchReactionEvent(event)) {
+      return;
+    }
+    log.info(
+      '(%s) (Room: "%s" %s) (Message: %s %s)',
+      event.getType(),
+      room?.name,
+      room?.roomId,
+      event.getSender(),
+      undefined,
+    );
     try {
-      if (!room) {
-        return;
-      }
       let initial = await client.roomInitialSync(room!.roomId, 1000);
       let eventList = (initial!.messages?.chunk || []) as DiscreteMatrixEvent[];
       let history: DiscreteMatrixEvent[] = constructHistory(eventList);
-      if (isPatchReactionEvent(event)) {
-        return await setTitle(
-          openai,
-          client,
-          room,
-          history,
-          aiBotUserId,
-          event,
-        );
-      }
-      return;
+      return await assistant.setTitle(room, history, event);
     } catch (e) {
       log.error(e);
       Sentry.captureException(e);
       return;
     }
+  });
+
+  //handle debug events
+  client.on(RoomEvent.Timeline, async function (event, room) {
+    if (event.event.origin_server_ts! < startTime) {
+      return;
+    }
+    if (event.getType() !== 'm.room.message') {
+      return;
+    }
+    if (!room) {
+      return;
+    }
+    let eventBody = event.getContent().body;
+    let isDebugEvent = eventBody.startsWith('debug:');
+    if (!isDebugEvent) {
+      return;
+    }
+    log.info(
+      '(%s) (Room: "%s" %s) (Message: %s %s)',
+      event.getType(),
+      room?.name,
+      room?.roomId,
+      event.getSender(),
+      eventBody,
+    );
+    return await assistant.handleDebugCommands(eventBody, room);
   });
 
   await client.startClient();
